% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_system.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate / Refresh Token}
\usage{
authenticate(x, ...)
}
\arguments{
\item{x}{A \code{HawkinAuth} object.}
}
\value{
An updated \code{HawkinAuth} object with a valid access token.
}
\description{
Internal S7 generic to exchange a Refresh Token for a new Access Token.
}
\keyword{internal}
