% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DateTimeParam}
\alias{DateTimeParam}
\title{Construct DateTime Parameters}
\usage{
DateTimeParam(param, value, sync)
}
\arguments{
\item{param}{The name of the parameter.}

\item{value}{The value of the parameter.}

\item{sync}{Logical indicating whether to use "sync" prefix.}
}
\value{
A string representing the constructed parameter.
}
\description{
This is an internal function used for constructing date-time parameters
for API requests within the package.
}
\keyword{internal}
