% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfDatetoChar}
\alias{dfDatetoChar}
\title{Convert Date-Time formats to Character Strings}
\usage{
dfDatetoChar(arg_df)
}
\arguments{
\item{arg_df}{A data frame containing flattened athlete test data.}
}
\value{
A data frame where the specified columns have been converted to nested lists.
}
\description{
This function takes a data frame of test trials and searches for any columns
with a date class. Then it will convert them to a character class.
}
\keyword{internal}
