% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfTests_expand}
\alias{dfTests_expand}
\title{Expand Comma-Separated Values To Nested Lists}
\usage{
dfTests_expand(arg_df)
}
\arguments{
\item{arg_df}{A data frame containing flattened athlete test data.}
}
\value{
A data frame where the specified columns have been converted to nested lists.
}
\description{
This function takes a data frame of athlete test output, which contains
comma-separated strings (test tag name, test tag id, test tag description,
athlete team, and athlete group), and expands them into nested lists.
}
\keyword{internal}
