% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_system.R
\docType{data}
\name{.hawkin_env}
\alias{.hawkin_env}
\title{Internal State Container}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.hawkin_env
}
\description{
This object is initialized in zzz.R during .onLoad().
It holds the active HawkinAuth connection.
}
\keyword{internal}
