% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{epochArgCheck}
\alias{epochArgCheck}
\title{Check for Proper EPOCH Times}
\usage{
epochArgCheck(arg.from, arg.to)
}
\arguments{
\item{from}{from argument provided into function.}

\item{to}{to argument provided into function.}
}
\value{
no object returned.
}
\description{
Check \code{from} and \code{to} parameters for proper class.
}
\keyword{internal}
