% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forcetime.R
\name{get_forcetime}
\alias{get_forcetime}
\title{Get Force-Time Data}
\usage{
get_forcetime(testId, ...)
}
\arguments{
\item{testId}{character. The unique identifier for the test trial.}

\item{...}{Optional arguments.
\itemize{
\item \code{profile}: A \code{HawkinAuth} object. If not provided, the active connection is used.
}}
}
\value{
A \code{HawkinForceTime} object.
}
\description{
Retrieves the raw force-time data for a specific test trial ID.
}
