% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get Test Metrics}
\usage{
get_metrics(testType = "all")
}
\arguments{
\item{testType}{Supply a value of type string. Must be canonical test Id, test type name, or test
name abbreviation.

Names | Abbreviations: "Countermovement Jump" | "CMJ", "Squat Jump" | "SJ",
"Isometric Test" | "ISO", "Drop Jump" | "DJ", "Free Run" | "FREE", "CMJ Rebound" | "CMJR",
"Multi Rebound" | "MR", "Weigh In" | "WI", "Drop Landing" | "DL", "TS Isometric Test" | "TSISO",
"TS Free Run" | "TSFREE"}
}
\value{
Response will be a data frame containing the tests metrics that are in the HD system. The parameter
\code{testType} can be used to filter and return only metrics of the specified type.

The returned data frame will follow the following schema:\tabular{lll}{
   \strong{Column Name} \tab \strong{Type} \tab \strong{Description} \cr
   \strong{canonicalTestTypeID} \tab \emph{chr} \tab Canonical Test Id \cr
   \strong{testTypeName} \tab \emph{chr} \tab Given Test Name \cr
   \strong{id} \tab \emph{chr} \tab camelCase Test Name \cr
   \strong{label} \tab \emph{chr} \tab Outward facing label or title \cr
   \strong{label_unit} \tab \emph{chr} \tab Outward facing label or title w/ unit of measure \cr
   \strong{header} \tab \emph{chr} \tab header of data frame output \cr
   \strong{units} \tab \emph{chr} \tab Unit of measure (if any) \cr
   \strong{description} \tab \emph{chr} \tab Description or definition of metric \cr
}
}
\description{
Get the metrics and ids for all the metrics in the system.
}
\examples{
\dontrun{
# This is an example of how the function would be called.

df_testsMetrics <- get_metrics(testType = "CMJ")

}

}
