% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{Get Teams}
\usage{
get_teams()
}
\value{
Response will be a data frame containing the teams that are in the organization.
Each team has the following variables:\tabular{lll}{
   \strong{Column Name} \tab \strong{Type} \tab \strong{Description} \cr
   \strong{id} \tab \emph{chr} \tab team's unique ID \cr
   \strong{name} \tab \emph{chr} \tab team's given name \cr
}
}
\description{
Get the team names and ids for all the teams in the org.
}
\examples{
\dontrun{
# This is an example of how the function would be called.

df_teams <- get_teams()

}

}
