% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests_ath.R
\name{get_tests_ath}
\alias{get_tests_ath}
\title{Get Test Trials By Athlete}
\usage{
get_tests_ath(athleteId, from, to)
}
\value{
Response will be a data frame containing the trials of the specified athlete within the time range (if specified)

id: \if{html}{\out{<chr>}} test's unique ID

name: \if{html}{\out{<chr>}} test's given name
}
\description{
Use this function to retrieve only tests of the specified athlete.
}
\examples{
\dontrun{
# This is an example of how the function would be called.


## Call for all tests from a specified athlete
df_cmj <- get_tests_ath(athleteId = "athleteId")


## Call for all tests within a specific time frame
dfFromTo <- get_tests_ath(athleteId = "athleteId", from = 1689958617, to = 1691207356)


## Call for all tests since a specific date
dfSince <- get_tests_ath("athleteId", from = 1689958617)

}

}
