% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests_team.R
\name{get_tests_team}
\alias{get_tests_team}
\title{Get Test Trials By Teams}
\usage{
get_tests_team(teamId, from, to)
}
\value{
Response will be a data frame containing the trials of the specified teams within the time range (if specified)

id: \if{html}{\out{<chr>}} test's unique ID

name: \if{html}{\out{<chr>}} test's given name
}
\description{
Use this function to retrieve only tests of the specified teams.
}
\examples{
\dontrun{
# This is an example of how the function would be called.


## Call for all tests by Group 1
dfteam1 <- get_tests_team(teamId = "team1")


## Call for all tests from Groups 1 & 2
dfteam_1_2 <- get_tests_team(teamId = paste0("team1","team2"))


## Call for all Group 1 tests since a specific date
df_team1_Since <- get_tests_team("team1", from = 1689958617)

}

}
