% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests_type.R
\name{get_tests_type}
\alias{get_tests_type}
\title{Get Test Trials By Test Type}
\usage{
get_tests_type(typeId, from, to)
}
\value{
Response will be a data frame containing the trials of the specified test type within the time range (if specified)

id: \if{html}{\out{<chr>}} test's unique ID

name: \if{html}{\out{<chr>}} test's given name
}
\description{
Use this function to retrieve only tests of the specified type.
}
\examples{
\dontrun{
# This is an example of how the function would be called.


## Call for all CMJ tests
df_cmj <- get_tests_type(typeId = "7nNduHeM5zETPjHxvm7s")


## Call for Free Run tests within a  specific time frame
df_free <- get_tests_type(typeId = "5pRSUQVSJVnxijpPMck3", from = 1689958617, to = 1691207356)


## Call for Squat Jump tests since a specific date
df_sjSince <- get_tests_type("QEG7m7DhYsD6BrcQ8pic", from = 1689958617)

}

}
