% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testTypes.R
\name{get_testTypes}
\alias{get_testTypes}
\title{Get Test Types}
\usage{
get_testTypes()
}
\value{
Response will be a data frame containing the tests that are in the HD system. Each test type has the following structure:

\strong{id} \if{html}{\out{<chr>}} test's unique ID

\strong{name} \if{html}{\out{<chr>}} test's given name
}
\description{
Use this function to retrieve the test names and IDs for the different tests in the Hawkin Dynamics system.
}
\examples{
\dontrun{
# This is an example of how the function would be called.

df_tests <- get_testTypes()

}

}
