% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_headers}
\alias{replace_headers}
\title{Clean Test Metric headers}
\usage{
replace_headers(df, typeId)
}
\arguments{
\item{df}{A data frame containing flattened test metric data.}

\item{testType}{the testType of the query}
}
\value{
A data frame with cleaned headers
#' @importFrom tidyselect everything
}
\description{
This function calls metric ids from the metric dictionary to clean metric name headers for test outputs.
}
\keyword{internal}
