% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncDB.R
\name{syncDB}
\alias{syncDB}
\title{Sync Database with Latest Test Data}
\usage{
syncDB(`file`, `includeInactive` =  FALSE, `newFile` = NULL)
}
\arguments{
\item{file}{A character string representing the path to the current database file. Supported file types include \code{csv}, \code{rds}, \code{xlsx}, \code{feather}, and \code{parquet}.}

\item{includeInactive}{A logical value indicating whether to include inactive tests in the sync. Defaults to \code{FALSE}.}

\item{newFile}{If NULL, updates will save to existing file path and overwrite previous version. Provide a new file path and name to create a new version of database and reserve original.}
}
\value{
The function saves the updated database to the specified file and logs the update process.
}
\description{
This function updates an existing test database by retrieving new tests and syncing updates
for previous tests. It reads the current database from the specified file and returns an updated version.
}
\details{
This function reads an existing database from the specified file and updates it by fetching new tests
and updating any changes to existing tests. It supports multiple file types and saves the updated database back to the original file.
Supported file formats include CSV, XLSX, Feather, Parquet, and RDS with optional compression.
}
\examples{
\dontrun{
# Sync an existing database and include inactive tests

syncDB("test_data.csv", includeInactive = TRUE, newFile = NULL)
}

}
