% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TagPrep}
\alias{TagPrep}
\title{Construct Test Type}
\usage{
TagPrep(arg.df)
}
\arguments{
\item{arg.df}{Data frame to be evaluated.}
}
\value{
data frame of test type information
}
\description{
Take testType section from data frame and prep for final data frame
}
\keyword{internal}
