% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TestIdCheck}
\alias{TestIdCheck}
\title{Check Test Id}
\usage{
TestIdCheck(arg_id)
}
\arguments{
\item{arg_id}{the testId argument provided in the function}
}
\value{
testId or error
}
\description{
Take testId argument and assess for correct format and validate before API call
}
\keyword{internal}
