% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{UpdateAthleteJSON}
\alias{UpdateAthleteJSON}
\title{Update Athlete Data Frame to JSON}
\usage{
UpdateAthleteJSON(arg_df)
}
\arguments{
\item{arg_df}{the athlete data frame argument provided in the function}
}
\value{
JSON string
}
\description{
Take the athlete data frame passed and convert to JSON for PUT method payload
}
\keyword{internal}
