% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_timestamp}
\alias{validate_timestamp}
\title{Validate time stamp arguments and allow for date character}
\usage{
validate_timestamp(x)
}
\arguments{
\item{x}{A data frame containing flattened athlete test data.}
}
\value{
An epoch time stamp
}
\description{
This function checks \code{from} and \code{to} arguments for valid inputs of epoch or
a date as a character string.
}
\keyword{internal}
