% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AthletePrep}
\alias{AthletePrep}
\title{Construct Athlete df}
\usage{
AthletePrep(arg_df)
}
\arguments{
\item{arg_df}{Data frame to be evaluated.}
}
\value{
data frame of test type information
}
\description{
Take athlete section from data frame and prep for final data frame
}
\keyword{internal}
