% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_token_validity}
\alias{check_token_validity}
\title{Check Authentication Validity}
\usage{
check_token_validity()
}
\description{
Takes the current token expiration and validates it
}
\keyword{internal}
