% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athletes.R
\name{get_athletes}
\alias{get_athletes}
\title{Get Athletes}
\usage{
get_athletes(inactive = FALSE)
}
\arguments{
\item{inactive}{FALSE by default to exclude inactive players in database. Set to TRUE if you want
inactive players included in the return.}
}
\value{
Response will be a data frame containing the athletes that match
this query. Each athlete has the following structure:

id: \if{html}{\out{<chr>}} athlete's unique ID

name: \if{html}{\out{<chr>}} athlete's given name (First Last)

active: \if{html}{\out{<logi>}} athlete is active (TRUE)

teams: \if{html}{\out{<chr>}} team ids separated by " | "

groups: \if{html}{\out{<chr>}} group ids separated by " | "
}
\description{
Use this function to retrieve all active athletes from your database. Inactive players will only be included if
\code{inactive} parameter is set to TRUE.
}
\examples{
\dontrun{
# This is an example of how the function would be called. If you only wish to call active players,
# you don't need to provide any parameters.

df_athletes <- get_athletes()

# If you want to include all athletes, including inactive athletes, include the optional
# `inactive` parameter.

df_inctvAthletes <- get_athletes(inactive = TRUE)
}


}
