% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athletes.R
\name{get_athletes}
\alias{get_athletes}
\title{Get Athletes}
\usage{
get_athletes(inactive = FALSE)
}
\arguments{
\item{inactive}{FALSE by default to exclude inactive players in database. Set to TRUE if you want
inactive players included in the return.}
}
\value{
Response will be a data frame containing the athletes that match this query.
Each athlete includes the following variables:

\strong{id}   \emph{chr}   athlete's unique ID

\strong{name}   \emph{chr}   athlete's given name (First Last)

\strong{active}   \emph{logi}   athlete is active (TRUE)

\strong{teams}   \emph{chr}   team ids separated by ","

\strong{groups}   \emph{chr}  group ids separated by ","

\strong{external}   \emph{chr} external ids as strings of "externalName:externalId" separated by ","
}
\description{
Get the athletes for an account. Inactive players will only be included if
\code{inactive} parameter is set to TRUE.
}
\examples{
\dontrun{
# This is an example of how the function would be called. If you only wish to call active players,
# you don't need to provide any parameters.

df_athletes <- get_athletes()

# If you want to include all athletes, including inactive athletes, include the optional
# `inactive` parameter.

df_wInactive <- get_athletes(inactive = TRUE)

}


}
