% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athletes.R
\name{get_athletes}
\alias{get_athletes}
\title{Get Athletes}
\usage{
get_athletes(includeInactive = FALSE)
}
\arguments{
\item{includeInactive}{FALSE by default to exclude inactive players in database. Set to TRUE if you want
inactive players included in the return.}
}
\value{
Response will be a data frame containing the athletes that match this query.
Each athlete includes the following variables:\tabular{lll}{
   \strong{Column Name} \tab \strong{Type} \tab \strong{Description} \cr
   \strong{id} \tab \emph{chr} \tab athlete's unique ID \cr
   \strong{name} \tab \emph{chr} \tab athlete's given name (First Last) \cr
   \strong{active} \tab \emph{bool} \tab athlete is active (TRUE) \cr
   \strong{teams} \tab \emph{chr} \tab team ids separated by "," \cr
   \strong{groups} \tab \emph{chr} \tab group ids separated by "," \cr
   \strong{external} \tab \emph{chr} \tab external properties will have a column of their name with the appropriate values for the athlete of \code{NA} if it does not apply \cr
}
}
\description{
Get the athletes for an account. Inactive players will only be included if
\code{includeInactive} parameter is set to TRUE.
}
\examples{
\dontrun{
# This is an example of how the function would be called. If you only wish to call active players,
# you don't need to provide any parameters.

df_athletes <- get_athletes()

# If you want to include all athletes, including inactive athletes, include the optional
# `includeInactive` parameter.

df_wInactive <- get_athletes( includeInactive = TRUE)

}

}
