% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forcetime.R
\name{get_forcetime}
\alias{get_forcetime}
\title{Get Force-Time Data}
\usage{
get_forcetime(testID)
}
\arguments{
\item{testID}{Give the unique test id of the trial you want to be called.}
}
\value{
Response will be a data frame containing the following:

\strong{time_s} (int): elapsed time in seconds, starting from end of identified quiet phase

\strong{force_right} (int): Force recorded from the RIGHT platform coinciding with time point from  \code{time_s}, measured in Newtons (N)

\strong{force_Left} (int): Force recorded from the LEFT platform coinciding with time point from  \code{time_s}, measured in Newtons (N)

\strong{force_combined} (int): Sum of forces from LEFT and RIGHT, coinciding with time point from  \code{time_s}, measured in Newtons (N)

\strong{velocity_m.s} (int): Calculated velocity of center of mass at time interval, measured in meters per second (m/s)

\strong{displacement_m} (int): Calculated displacement of center of mass at time interval, measured in meters (m)

\strong{power_w} (int): Calculated power of mass at time interval, measured in watts (W)
}
\description{
Use this function to retrieve the raw recorded force-time data from a trial. This includes both left, right and combined force data at 1000hz (per millisecond).
Calculated velocity, displacement, and power at each time interval will also be included.
}
\examples{
\dontrun{
# This is an example of how the function would be called.

df_forcetime <- get_forcetime( testID = `stringId` )

}

}
