% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Get Groups}
\usage{
get_groups()
}
\value{
Response will be a data frame containing the groups that are in the organization.Each group has the following structure:

id: \if{html}{\out{<chr>}} group's unique ID

name: \if{html}{\out{<chr>}} group's given name
}
\description{
Use this function to retrieve the group names and IDs in the organization.
}
\examples{
\dontrun{
# This is an example of how the function would be called.

df_groups <- get_groups()

}

}
