% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get Tags}
\usage{
get_tags()
}
\value{
Response will be a data frame containing the tags that are in the organization.
Each tag has the following variables:

\strong{id}   \emph{chr}   tag's unique ID

\strong{name}   \emph{chr}   tag's given name

\strong{description}   \emph{chr}   description of tag provided by user
}
\description{
Get the tag names and ids for all the tags in the system.
}
\examples{
\dontrun{
# This is an example of how the function would be called.

df_tags <- get_tags()

}

}
