% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get Tags}
\usage{
get_tags()
}
\value{
Response will be a data frame containing the tags that are in the organization.
Each tag has the following variables:\tabular{lll}{
   \strong{Column Name} \tab \strong{Type} \tab \strong{Description} \cr
   \strong{id} \tab \emph{chr} \tab tag's unique ID \cr
   \strong{name} \tab \emph{chr} \tab tag's given name \cr
   \strong{description} \tab \emph{chr} \tab description of tag provided by user \cr
}
}
\description{
Get the tag names and ids for all the tags in the system.
}
\examples{
\dontrun{
# This is an example of how the function would be called.

df_tags <- get_tags()

}

}
