% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{Get Teams}
\usage{
get_teams()
}
\value{
Response will be a data frame containing the teams that are in the organization.Each team has the following structure:

id: \if{html}{\out{<chr>}} team's unique ID

name: \if{html}{\out{<chr>}} team's given name
}
\description{
Use this function to retrieve the team names and IDs in the organization.
}
\examples{
\dontrun{
# This is an example of how the function would be called.

df_teams <- get_teams()

}

}
