% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests.R
\name{get_tests}
\alias{get_tests}
\title{Get All Tests or Sync Tests}
\usage{
get_tests(from, to, sync = FALSE)
}
\value{
Response will be a data frame containing the trials within the time range (if specified)

id: \if{html}{\out{<chr>}} test's unique ID

name: \if{html}{\out{<chr>}} test's given name
}
\description{
Use this function to retrieve all tests or sync tests from your database. You can specify a time frame from, or to, which the tests should come (or be synced)
}
\examples{
\dontrun{
# This is an example of how the function would be called.


## Call for all tests
dfAllTests <- get_tests()


## Call for all tests within a specific time frame
dfFromTo <- get_tests(from = 1689958617, to = 1691207356)


## Call for all new tests since a specific date, or any tests that have been updated/changed since that date
dfSync <- get_tests(from = 1689958617, sync=TRUE)

}

}
