% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests_ath.R
\name{get_tests_ath}
\alias{get_tests_ath}
\title{Get Test Trials By Athlete}
\usage{
get_tests_ath(athleteId, from, to)
}
\arguments{
\item{athleteId}{Supply an athlete’s id to receive tests for a specific athlete}

\item{from}{Optionally supply a time (Unix timestamp) you want the tests from. If you do not
supply this value you will receive every test. This parameter is best suited for bulk exports of
historical data}

\item{to}{Optionally supply a time (Unix timestamp) you want the tests to. If you do not
supply this value you will receive every test from the beginning of time or the optionally
supplied \code{from} parameter. This parameter is best suited for bulk exports of historical data.}
}
\value{
Response will be a data frame containing the trials from the specified team and within the time range (if specified).

\strong{id}   \emph{str}   Test trial unique ID

\strong{timestamp}   \emph{int}   UNIX timestamp of trial

\strong{segment}   \emph{chr}   Description of the test type and trial number of the session (testType:trialNo)

\strong{testType.id}   \emph{chr}   Id of the test type of the trial

\strong{testType.name}   \emph{chr}   Name of the test type of the trial

\strong{testType.canonicalId}   \emph{chr}   Canonical Id of the test type of the trial

\strong{athlete.id}   \emph{chr}   Unique Id of the athlete

\strong{athlete.name}   \emph{chr}   Athlete given name

\strong{athlete.active}   \emph{logi}   The athlete is active

\strong{athlete.teams}   \emph{list}   List containing Ids of each team the athlete is on

\strong{athlete.groups}   \emph{list}   List containing Ids of each group the athlete is in

All metrics from each test type are included as the remaining variables.
If a trial does not have data for a variable it is returned NA.
}
\description{
Get only tests of the specified athlete for an account.
}
\examples{
\dontrun{
# This is an example of how the function would be called.


## Call for all tests from a specified athlete
df_cmj <- get_tests_ath(athleteId = "athleteId")


## Call for all tests within a specific time frame
dfFromTo <- get_tests_ath(athleteId = "athleteId", from = 1689958617, to = 1691207356)


## Call for all tests since a specific date
dfSince <- get_tests_ath("athleteId", from = 1689958617)

}

}
