% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests_group.R
\name{get_tests_group}
\alias{get_tests_group}
\title{Get Test Trials By Groups}
\usage{
get_tests_group(groupId, from, to)
}
\value{
Response will be a data frame containing the trials of the specified groups within the time range (if specified)

id: \if{html}{\out{<chr>}} test's unique ID

name: \if{html}{\out{<chr>}} test's given name
}
\description{
Use this function to retrieve only tests of the specified groups.
}
\examples{
\dontrun{
# This is an example of how the function would be called.


## Call for all tests by Group 1
dfGroup1 <- get_tests_group(groupId = "group1")


## Call for all tests from Groups 1 & 2
dfGroups_1_2 <- get_tests_group(groupId = paste0("group1","group2"))


## Call for all Group 1 tests since a specific date
df_Group1_Since <- get_tests_group("group1", from = 1689958617)

}

}
