% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests_type.R
\name{get_tests_type}
\alias{get_tests_type}
\title{Get Test Trials By Test Type}
\usage{
get_tests_type(typeId, from, to, sync)
}
\arguments{
\item{typeId}{Supply a test type id to only retrieve tests of that type.}

\item{from}{Optionally supply a time (Unix timestamp) you want the tests from. If you do not
supply this value you will receive every test. This parameter is best suited for bulk exports of
historical data.}

\item{to}{Optionally supply a time (Unix timestamp) you want the tests to. If you do not
supply this value you will receive every test from the beginning of time or the optionally
supplied \code{from} parameter. This parameter is best suited for bulk exports of historical data.}

\item{sync}{The result set will include updated and newly created tests. This parameter is best
suited to keep your database in sync with the Hawkin database. If you do not supply this value
you will receive every test.}
}
\value{
Response will be a data frame containing the trials of the specified type and within the time range (if specified).

\strong{id}   \emph{str}   Test trial unique ID

\strong{timestamp}   \emph{int}   UNIX timestamp of trial

\strong{segment}   \emph{chr}   Description of the test type and trial number of the session (testType:trialNo)

\strong{testType.id}   \emph{chr}   Id of the test type of the trial

\strong{testType.name}   \emph{chr}   Name of the test type of the trial

\strong{testType.canonicalId}   \emph{chr}   Canonical Id of the test type of the trial

\strong{athlete.id}   \emph{chr}   Unique Id of the athlete

\strong{athlete.name}   \emph{chr}   Athlete given name

\strong{athlete.active}   \emph{logi}   The athlete is active

\strong{athlete.teams}   \emph{list}   List containing Ids of each team the athlete is on

\strong{athlete.groups}   \emph{list}   List containing Ids of each group the athlete is in

Only metrics of the given test type are included as the remaining variables.
}
\description{
Get only tests of the specified type for an account.
}
\examples{
\dontrun{
# This is an example of how the function would be called.


## Call for all CMJ tests
df_cmj <- get_tests_type(typeId = "7nNduHeM5zETPjHxvm7s")


## Call for Free Run tests within a  specific time frame
df_free <- get_tests_type(typeId = "5pRSUQVSJVnxijpPMck3", from = 1689958617, to = 1691207356)


## Call for Squat Jump tests since a specific date
df_sjSince <- get_tests_type("QEG7m7DhYsD6BrcQ8pic", from = 1689958617)

}

}
