% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tests_type.R
\name{get_tests_type}
\alias{get_tests_type}
\title{Get Test Trials By Test Type}
\usage{
get_tests_type(typeId, from = NULL, to = NULL, sync = FALSE, includeInactive = FALSE)
}
\arguments{
\item{typeId}{Supply a value of type string. Must be canonical test Id, test type name, or test
name abbreviation.

Names | Abbreviations: "Countermovement Jump" | "CMJ", "Squat Jump" | "SJ",
"Isometric Test" | "ISO", "Drop Jump" | "DJ", "Free Run" | "FREE", "CMJ Rebound" | "CMJR",
"Multi Rebound" | "MR", "Weigh In" | "WI", "Drop Landing" | "DL"}

\item{from}{Optionally supply a time (Unix time stamp) you want the tests from. If you do not
supply this value you will receive every test. This parameter is best suited for bulk exports of
historical data.}

\item{to}{Optionally supply a time (Unix time stamp) you want the tests to. If you do not
supply this value you will receive every test from the beginning of time or the optionally
supplied \code{from} parameter. This parameter is best suited for bulk exports of historical data.}

\item{sync}{The result set will include updated and newly created tests. This parameter is best
suited to keep your database in sync with the Hawkin database. If you do not supply this value
you will receive every test.}

\item{includeInactive}{There was a change to the default API configuration to reflect the majority of
users API configuration. Inactive tests or tests where \code{active = false} are returned in these
configuration. Be default, \code{includeInactive} is set to \code{FALSE}. To return all tests, including disabled
trials, set \code{includeInactive} to \code{TRUE}.}
}
\value{
Response will be a data frame containing the trials within the time range (if specified).\tabular{lll}{
   \strong{Column Name} \tab \strong{Type} \tab \strong{Description} \cr
   \strong{id} \tab \emph{str} \tab Test trial unique ID \cr
   \strong{active} \tab \emph{logi} \tab The trial is active and not disabled \cr
   \strong{timestamp} \tab \emph{int} \tab UNIX time stamp of trial \cr
   \strong{segment} \tab \emph{chr} \tab Description of the test type and trial number of the session (testType:trialNo) \cr
   \strong{test_type_id} \tab \emph{chr} \tab Id of the test type of the trial \cr
   \strong{test_type_name} \tab \emph{chr} \tab Name of the test type of the trial \cr
   \strong{test_type_canonicalId} \tab \emph{chr} \tab Canonical Id of the test type of the trial \cr
   \strong{test_type_tag_ids} \tab \emph{chr} \tab String of Ids associated with tags used during the test trial \cr
   \strong{test_type_tag_names} \tab \emph{chr} \tab String of names of tags used during the test trial \cr
   \strong{test_type_tag_desc} \tab \emph{chr} \tab String of descriptions of tags used during the test trial \cr
   \strong{athlete_id} \tab \emph{chr} \tab Unique Id of the athlete \cr
   \strong{athlete_name} \tab \emph{chr} \tab Athlete given name \cr
   \strong{athlete_active} \tab \emph{logi} \tab The athlete is active \cr
   \strong{athlete_teams} \tab \emph{list} \tab List containing Ids of each team the athlete is on \cr
   \strong{athlete_groups} \tab \emph{list} \tab List containing Ids of each group the athlete is in \cr
}


Only metrics of the given test type are included as the remaining variables.
}
\description{
\strong{Deprecated}: Use \code{get_tests} instead, which has been expanded to handle all requests.

Get only tests of the specified type for an account.
}
\examples{
\dontrun{
# This is an example of how the function would be called.


## Call for all CMJ tests
df_cmj <- get_tests_type(typeId = "7nNduHeM5zETPjHxvm7s")


## Call for Free Run tests within a  specific time frame
df_free <- get_tests_type(typeId = "5pRSUQVSJVnxijpPMck3", from = 1689958617, to = 1691207356)


## Call for Squat Jump tests since a specific date
df_sjSince <- get_tests_type("QEG7m7DhYsD6BrcQ8pic", from = 1689958617)

}

}
\seealso{
\code{get_tests}
}
