% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testTypes.R
\name{get_testTypes}
\alias{get_testTypes}
\title{Get Test Types}
\usage{
get_testTypes()
}
\value{
Response will be a data frame containing the tests that are in the HD system.
Each test type includes the following variables:\tabular{lll}{
   \strong{Column Name} \tab \strong{Type} \tab \strong{Description} \cr
   \strong{canonicalId} \tab \emph{chr} \tab test's unique canonical ID \cr
   \strong{name} \tab \emph{chr} \tab test's given name \cr
   \strong{abbreviation} \tab \emph{chr} \tab test given name abbreviation \cr
}
}
\description{
Get the test type names and ids for all the test types in the system.
}
\examples{
\dontrun{
# This is an example of how the function would be called.

df_tests <- get_testTypes()

}

}
