% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_system.R
\name{hd_auth_reset}
\alias{hd_auth_reset}
\title{Remove Hawkin Credentials}
\usage{
hd_auth_reset(profile = "default")
}
\arguments{
\item{profile}{character. The name of the profile to remove.}
}
\description{
Deletes a stored Refresh Token from the system keychain.
}
