% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ParamValidation}
\alias{ParamValidation}
\title{Validate GetTest Parameters}
\usage{
ParamValidation(
  arg_athleteId = NULL,
  arg_testTypeId = NULL,
  arg_teamId = NULL,
  arg_groupId = NULL
)
}
\value{
no object returned.
}
\description{
Check \code{athleteId}, \code{testTypeId}, \code{teamId}, and \code{groupId} parameters.
}
\keyword{internal}
