% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testIdCheck}
\alias{testIdCheck}
\title{Check Test Id}
\usage{
testIdCheck(arg.id)
}
\arguments{
\item{arg.id}{the testId argument provided in the function}
}
\value{
testId or error
}
\description{
Take testId argument and assess for correct format and validate before API call
}
\keyword{internal}
