% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_athletes.R
\name{update_athletes}
\alias{update_athletes}
\title{Update Athletes}
\usage{
update_athletes(athleteData)
}
\arguments{
\item{athleteData}{Provide a data frame of the athlete or athletes to be updated.}
}
\value{
If successful, a confirmation message will be printed with the number of successful athletes created.
If there are failures, a data frame containing the athletes that failed to be created will be returned with columns:\tabular{lll}{
   \strong{Column Name} \tab \strong{Type} \tab \strong{Description} \cr
   \strong{reason} \tab \emph{chr} \tab Reason for failed creation \cr
   \strong{name} \tab \emph{chr} \tab Athlete's given name (First Last) \cr
}
}
\description{
Update an athlete or athletes for an account. Bulk update up to 500 athletes at a time.
}
\details{
The data frame passed as the argument must use the following schema:\tabular{llll}{
   \strong{Column Name} \tab \strong{Type} \tab \strong{Inclusion} \tab \strong{Description} \cr
   \strong{id} \tab \emph{chr} \tab \strong{REQUIRED} \tab athlete's Hawkin Dynamics unique ID \cr
   \strong{name} \tab \emph{chr} \tab \emph{optional} \tab athlete's given name (First Last) \cr
   \strong{image} \tab \emph{chr} \tab \emph{optional} \tab URL path to image. \code{default = null} \cr
   \strong{active} \tab \emph{logi} \tab \emph{optional} \tab athlete is active (TRUE). \code{default = null} \cr
   \strong{teams} \tab \emph{list} \tab \emph{optional} \tab a single team id as a string or list of team ids. \verb{default = [defaultTeamId]} \cr
   \strong{groups} \tab \emph{list} \tab \emph{optional} \tab a single group id as a string or list of group ids. \verb{default = []} \cr
   \strong{external property} \tab \emph{chr} \tab \emph{optional} \tab External properties can be added by adding any additional columns of equal length. The name of the column will become the external property name, and the row value will become the external property value. Use "lowercase" or "snake_case" styles for column names. \cr
}


\emph{If optional fields are not present in an update request, those properties will be left unchanged. However, when updating external properties, custom properties that are not present will be removed.}
}
\examples{
\dontrun{
# Example data frame following the required schema
df <- data.frame(
  id = c("uniqueAthleteIdOne", "uniqueAthleteIdTwo"),
  name = c("John Doe", "Jane Smith"),
  image = c("http://example.com/johndoe.jpg", "http://example.com/janesmith.jpg"),
  active = c(TRUE, FALSE),
  teams = c("team1", c("team2", "team3"))),
  groups = c(NULL, "group1")),
  external_property = c("value1", "value2")
)

# Update athletes using the example data frame
update_athletes(athleteData = df)
}

}
